move spectral_indices files
