+++ demos/a_hsi_img_for_VI_demo.mat
