+++ demos/spectral_indicies_demo_results/CAI.png
