+++ demos/spectral_indicies_demo_results/CARI.png
