+++ demos/spectral_indicies_demo_results/CIrededge.png
