+++ demos/spectral_indicies_demo_results/CRI1.png
