+++ demos/spectral_indicies_demo_results/CRI2.png
