+++ demos/spectral_indicies_demo_results/EVI.png
