+++ demos/spectral_indicies_demo_results/MCARI.png
