+++ demos/spectral_indicies_demo_results/MTCI.png
