+++ demos/spectral_indicies_demo_results/NDII.png
