+++ demos/spectral_indicies_demo_results/NDLI.png
