+++ demos/spectral_indicies_demo_results/NDNI.png
