+++ demos/spectral_indicies_demo_results/NDRE.png
