+++ demos/spectral_indicies_demo_results/NDVI.png
