+++ demos/spectral_indicies_demo_results/NDWI.png
