+++ demos/spectral_indicies_demo_results/PRI.png
