+++ demos/spectral_indicies_demo_results/PSRI.png
