+++ demos/spectral_indicies_demo_results/PSSR1.png
