+++ demos/spectral_indicies_demo_results/PSSR2.png
