+++ demos/spectral_indicies_demo_results/PSSR3.png
