+++ demos/spectral_indicies_demo_results/RVSI.png
