+++ demos/spectral_indicies_demo_results/SIPI.png
