+++ demos/spectral_indicies_demo_results/SR.png
