+++ demos/spectral_indicies_demo_results/VARI.png
