+++ demos/spectral_indicies_demo_results/VIgreen.png
