+++ demos/spectral_indicies_demo_results/WBI.png
