+++ demos/spectral_indicies_demo_results/WDVI.png
