+++ hsi_toolkit/__init__.py
@@ -1,3 +1,4 @@
 from anomaly_detectors import *
 from endmember_extraction import *
 from signature_detectors import *
+from spectral_indices import *
